src/flare_ai_rag/ai/__init__.py
from .base import AsyncBaseClient, BaseClient, BaseAIProvider, ModelResponse
from google.generativeai.embedding import EmbeddingTaskType
from .gemini import GeminiEmbedding, GeminiProvider
from .model import Model
from .openrouter import OpenRouterClient

__all__ = [
    "AsyncBaseClient",
    "BaseClient",
    "BaseAIProvider",
    "ModelResponse",
    "EmbeddingTaskType",
    "GeminiEmbedding",
    "GeminiProvider",
    "Model",
    "OpenRouterClient",
]
